requirements.txt
numpy
Pillow
Flask
python-dotenv